Refactor BLE configuration and add device configuration files; update debug level and include necessary headers
