+++ platformio.ini
@@ -61,7 +61,7 @@ lib_deps =
 	fastled/FastLED@^3.9.14
 build_flags = 
 	-std=gnu++17
-	-DARDUINO_DEBUG_LEVEL=6
+	-DCORE_DEBUG_LEVEL=3
 
 [env:adafruit_metro_esp32s3]
 board = adafruit_metro_esp32s3
